Check if dir exists before removing. Add README

+++ build_and_run.sh
@@ -7,7 +7,9 @@ geant_version="10.2.1"
 lab_name="B1"
 geant_dir="/Users/oleksa/univ/Geant4-${geant_version}-Darwin"
 
-rm -r "${cur_dir}/${lab_name}-build"
+if [ -d "${cur_dir}/${lab_name}-build" ]; then
+    rm -r "${cur_dir}/${lab_name}-build"
+fi
 mkdir "${cur_dir}/${lab_name}-build"
 cd "${cur_dir}/${lab_name}-build"
 cmake -DGeant4_DIR="${geant_dir}/lib/Geant4-${geant_version}" "../${lab_name}"
